requirements.txt
feedparser
atproto
Mastodon.py
requests
beautifulsoup4